Fix: stats type mismatches, dead fields, and redundant counters
- Fix int/uint32_t -> uint64_t for stat fields in st40_rx and st41_rx
- Remove 9 wrong-domain RX fields from st41_tx_user_stats (never set)
- Remove stale stat_last_time copy in TX fastmetadata stat dump
- Remove redundant stat_pkts_out_of_order_per_port internal counter
  from all 4 RX session types (already tracked via port_user_stats)
- Add per-port out-of-order breakdown for RX video sessions
- Use snapshot-based deltas for per-port OOO logging in all sessions
- Fix TX video stat_pkts_burst: print before reset (was silently zeroed)
- Fix format specifiers in RX fastmetadata stat dump to match uint64_t

+++ include/st40_api.h
@@ -477,7 +477,7 @@ struct st40_rx_user_stats {
   uint64_t stat_pkts_enqueue_fail;
   uint64_t stat_interlace_first_field;
   uint64_t stat_interlace_second_field;
-  int stat_pkts_wrong_interlace_dropped;
+  uint64_t stat_pkts_wrong_interlace_dropped;
 };
 
 /**
